Rename config getters to match reality
